@@ -1,69 +1,64 @@
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Localization;
 using Microsoft.AspNetCore.Mvc.Razor;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
+
 using System.Globalization;
 
-namespace ASPNETCoreLocalization
+namespace ASPNETCoreLocalization;
+
+public class Startup
 {
-    public class Startup
+    public Startup(IConfiguration configuration)
     {
-        public Startup(IConfiguration configuration)
+        Configuration = configuration;
+    }
+
+    public IConfiguration Configuration { get; }
+
+    // This method gets called by the runtime. Use this method to add services to the container.
+    public void ConfigureServices(IServiceCollection services)
+    {
+        services.AddLocalization(options => options.ResourcesPath = "Resources");
+
+        services.AddRazorPages()
+            .AddViewLocalization(LanguageViewLocationExpanderFormat.SubFolder)
+            .AddDataAnnotationsLocalization();
+    }
+
+    // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+        if (env.IsDevelopment())
         {
-            Configuration = configuration;
+            app.UseDeveloperExceptionPage();
+        }
+        else
+        {
+            app.UseExceptionHandler("/Error");
+            // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+            app.UseHsts();
         }
 
-        public IConfiguration Configuration { get; }
+        CultureInfo[] supportedCultures = { new("en-US"), new("en"), new("de-AT"), new("de") };
 
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
+        RequestLocalizationOptions localizationOptions = new()
         {
-            services.AddLocalization(options => options.ResourcesPath = "Resources");
+            DefaultRequestCulture = new RequestCulture(new CultureInfo("en-US")),
+            SupportedCultures = supportedCultures,
+            SupportedUICultures = supportedCultures
+        };
 
-            services.AddRazorPages()
-                .AddViewLocalization(LanguageViewLocationExpanderFormat.SubFolder)
-                .AddDataAnnotationsLocalization();
-        }
+        app.UseRequestLocalization(localizationOptions);
 
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+        app.UseHttpsRedirection();
+        app.UseStaticFiles();
+
+        app.UseRouting();
+
+        app.UseAuthorization();
+
+        app.UseEndpoints(endpoints =>
         {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            else
-            {
-                app.UseExceptionHandler("/Error");
-                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-                app.UseHsts();
-            }
-
-            CultureInfo[] supportedCultures = { new("en-US"), new("en"), new("de-AT"), new("de") };
-
-            RequestLocalizationOptions localizationOptions = new()
-            {
-                DefaultRequestCulture = new RequestCulture(new CultureInfo("en-US")),
-                SupportedCultures = supportedCultures,
-                SupportedUICultures = supportedCultures
-            };
-
-            app.UseRequestLocalization(localizationOptions);
-
-            app.UseHttpsRedirection();
-            app.UseStaticFiles();
-
-            app.UseRouting();
-
-            app.UseAuthorization();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapRazorPages();
-            });
-        }
+            endpoints.MapRazorPages();
+        });
     }
 }
